added more variables in orderscontext from orders file to be accessed for both orders and ordersmgmt
